Split PlantingSiteStoreReadTest
This test was getting long and with upcoming changes, it would have gotten even
longer. Split it up into smaller top-level test classes.

No changes to the actual test code here, just moving the inner classes to new
top-level classes.
